% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.fit.R
\name{DiagAver}
\alias{DiagAver}
\title{Diagonal averaging}
\usage{
DiagAver(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A vector.
}
\description{
Diagonal averaging
}
\examples{
PC1<- Group(AirPassengers[1:20], L=10, groups=1)
DiagAver(PC1)
}
