% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.fit.R
\name{Group}
\alias{Group}
\title{Grouping}
\usage{
Group(y, L, groups)
}
\arguments{
\item{y}{A time series or a vector.}

\item{L}{The window length.}

\item{groups}{The number of eigentriples for reconstruction.}
}
\value{
A matrix with the approximation.
}
\description{
Grouping
}
\examples{
Group(AirPassengers[1:20], L=10, groups=5)
}
