% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel_matrix.R
\name{hankel_vector}
\alias{hankel_vector}
\title{Transform a hankel matrix in a vector}
\usage{
hankel_vector(H)
}
\arguments{
\item{H}{A hankel matrix.}
}
\value{
A vector.
}
\description{
Transform a hankel matrix in a vector
}
\examples{
y <- datasets::AirPassengers
L <- 12
H <- hankel_matrix(y, L)
hankel_vector(H)
}
