% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.hts-package.R
\docType{package}
\name{ssa.hts-package}
\alias{ssa.hts}
\alias{ssa.hts-package}
\title{ssa.hts: Hierarchical Time Series Forecasting using Singular Spectrum Analysis}
\description{
The singular spectrum analysis will be used to create an extra hierarchy to improve hierarchical time series forecasting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/numbats/assignment-1-package-creation-paulocanas}
  \item \url{https://paulocanas.github.io/ssa.hts/}
  \item Report bugs at \url{https://github.com/numbats/assignment-1-package-creation-paulocanas/issues}
}

}
\author{
\strong{Maintainer}: Paulo Canas Rodrigues \email{paulocanas@gmail.com} (\href{https://orcid.org/0000-0002-1248-9910}{ORCID})

}
\keyword{internal}
