% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.fit.R
\name{ssa}
\alias{ssa}
\title{Singular spectrum analysis}
\usage{
ssa(y, L, groups)
}
\arguments{
\item{y}{A time series or a vector.}

\item{L}{The window length.}

\item{groups}{The number of eigentriples for reconstruction.}
}
\value{
A list with the approximation and the residual.
}
\description{
Singular spectrum analysis
}
\examples{
s<- ssa(AirPassengers[1:20], L=10, groups=5)
s$approximation
s$residual
AirPassengers[1:20] - (s$approximation + s$residual)
}
