% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.fit.R
\name{SVD}
\alias{SVD}
\title{Singular value decomposition}
\usage{
SVD(y, L)
}
\arguments{
\item{y}{A time series or a vector.}

\item{L}{The window length.}
}
\value{
A list with the singular value decomposition.
}
\description{
Singular value decomposition
}
\examples{
SVD(AirPassengers[1:20], L=10)
}
