% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.fit.R
\name{UniHankel}
\alias{UniHankel}
\title{Trajectory matrix}
\usage{
UniHankel(y, L)
}
\arguments{
\item{y}{A time series or a vector.}

\item{L}{The window length.}
}
\value{
A hankel matrix.
}
\description{
Trajectory matrix
}
\examples{
UniHankel(AirPassengers[1:20], L=10)
}
